Add database creation and deletion scripts
These scripts should be used to create the database on your machine.
create_database.bat makes a new file called create.sql, which is just a
concatenation of each sql file in the proper order for creation of the
database. delete_database.bat simply runs delete.sql, which just drops
each table and new type in the proper order.