Streamline database creation and deletion process
New scripts to create the initial database and delete the entire database were
added. These should only need to be used once, the add_tables and remove_tables
scripts should be used whenever you want to recreate the tables in the database
or drop all tables in the database.

Some minor changes were made to the tables in school and year_info.