Add Box Score Scraper
This scraper gets the box scores for a given year and division. It makes three
csv files, for hitting, pitching, and fielding. It uses the game id that comes
from the game info scraper, since in 2019 a contest id was added to the main
game page of every game, which could not be used to get each box score type. The
game info scraper gets the real game id of every game.